Remove Twine from the build dependencies.
Twine is needed only to upload the built package to PyPI, and is not used when
building the package. This is described by Jason Andryuk in
<URL: https://pagure.io/python-daemon/issue/55>.

Now that we have Twine declared as a ‘devel’ requirement, and install those as
part of the Make rules for invoking Twine, we remove it from the build
requirements.